Refactor form use to allow for form auto-generation
Drops manual modifications to anki_simulator_dialog form in favor
of setting the graph widget up at SimulatorDialog initialization

The form can now be compiled with `aab ui`.
