+++ designer/anki_simulator_dialog.ui
@@ -592,16 +592,6 @@
      </layout>
     </widget>
    </item>
-   <item>
-    <widget class="QWidget" name="simulationGraph" native="true">
-     <property name="minimumSize">
-      <size>
-       <width>0</width>
-       <height>227</height>
-      </size>
-     </property>
-    </widget>
-   </item>
   </layout>
  </widget>
  <tabstops>
