+++ src/anki_simulator/gui/chart/LICENSE.md
