+++ src/anki_simulator/gui/graph.html
